refactor: Address 임베디드 클래스에서 도메인으로 리팩터링
DB에 중복되는 컬럼이 너무 많아짐에 따라..
