+++ src/main/java/com/example/temp/baseball/domain/Stadium.java
@@ -1,9 +1,10 @@
 package com.example.temp.baseball.domain;
 
-import com.example.temp.common.entity.Address;
-import jakarta.persistence.Embedded;
+import com.example.temp.geo.entity.Address;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class Stadium {
     @Id
     private Long id;
-    @Embedded
+    @ManyToOne(fetch = FetchType.LAZY)
     private Address address;
     private String aliasAddr;
