+++ src/main/java/com/example/temp/fanpool/domain/Fanpool.java
@@ -1,13 +1,16 @@
 package com.example.temp.fanpool.domain;
 
 import com.example.temp.baseball.domain.Game;
-import com.example.temp.common.entity.Address;
 import com.example.temp.common.entity.BaseTimeEntity;
 import com.example.temp.common.util.IdUtil;
 import com.example.temp.fanpool.domain.value.FanpoolType;
 import com.example.temp.fanpool.dto.CreateFanpoolRequest;
+import com.example.temp.geo.entity.Address;
 import com.example.temp.user.domain.User;
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -27,8 +30,8 @@ public class Fanpool extends BaseTimeEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private Game game;
 
-    @Embedded
-    private Address place;
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Address departFrom;
     private LocalDateTime departAt;
     private Integer numberOfPeople;
     private Integer currentNumberOfPeople;
@@ -42,7 +45,7 @@ public class Fanpool extends BaseTimeEntity {
         ret.hostUser = hostUser;
         ret.departAt = request.getDepartAt();
         ret.numberOfPeople = request.getNumberOfPeople();
-        ret.currentNumberOfPeople = 0;
+        ret.currentNumberOfPeople = 1;
         ret.fanpoolType = request.getFanpoolType();
         ret.memo = request.getMemo();
         return ret;
