+++ src/main/java/com/example/temp/geo/entity/Address.java
@@ -1,6 +1,9 @@
-package com.example.temp.common.entity;
+package com.example.temp.geo.entity;
 
-import jakarta.persistence.Embeddable;
+import com.example.temp.common.entity.BaseTimeEntity;
+import com.example.temp.common.util.IdUtil;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,10 +13,14 @@ import lombok.NoArgsConstructor;
  * 참고 - 행정안전부 주소기반산업지원서비스
  * https://business.juso.go.kr/addrlink/elctrnMapProvd/geoDBDwldPubList.do?cPath=99JD
  */
-@Embeddable
+
+@Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class Address {
+public class Address extends BaseTimeEntity {
+    @Id
+    private Long id;
+
     private String fullText;    // 전체 주소
     private String zipNo;
     private String dongCd;      // 법정동 코드(6) = 읍면동일련번호(2) + 읍면동구분(1, 0:읍면, 1:동, 2:미부여) + 법정동기준읍면동코드(3)
@@ -21,6 +28,7 @@ public class Address {
     private String sigungu;     // 시군구명
     private String dong;        // 법정동명
     private String eupmyun;     // 법정읍면명
+
     private String doroCd;      // 도로명 코드 = 시군구번호(5) + 도로명번호(7)
     private String doro;        // 도로명
     private String mainBuildNo; // 본 건물번호
@@ -33,6 +41,7 @@ public class Address {
 
     @Builder
     public Address(String fullText, String zipNo, String dongCd, String sido, String sigungu, String dong, String eupmyun, String doroCd, String doro, String mainBuildNo, String subBuildNo, String buildName, String detailAddr, String x, String y) {
+        this.id = IdUtil.create();
         this.fullText = fullText;
         this.zipNo = zipNo;
         this.dongCd = dongCd;
