+++ src/main/java/com/example/temp/geo/entity/AddressRepository.java
@@ -1,0 +1,7 @@
+package com.example.temp.geo.entity;
+
+import org.springframework.data.repository.Repository;
+
+public interface AddressRepository extends Repository<Address, Long> {
+    Address save(Address address);
+}
