+++ src/main/java/com/example/temp/geo/entity/AuthenticatedLocation.java
@@ -1,11 +1,8 @@
 package com.example.temp.geo.entity;
 
-import com.example.temp.common.entity.Address;
 import com.example.temp.common.entity.BaseTimeEntity;
 import com.example.temp.common.util.IdUtil;
-import jakarta.persistence.Embedded;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,7 +14,7 @@ public class AuthenticatedLocation extends BaseTimeEntity {
     @Id
     private Long id;
 
-    @Embedded
+    @ManyToOne(fetch = FetchType.LAZY)
     private Address address;
 
     public AuthenticatedLocation(Address address) {
