+++ src/main/java/com/example/temp/geo/service/impl/Coord2AddressClient.java
@@ -1,20 +1,19 @@
 package com.example.temp.geo.service.impl;
 
-import com.example.temp.common.entity.Address;
 import com.example.temp.common.exception.CustomException;
 import com.example.temp.geo.dto.VworldCoord2AddressResponse;
+import com.example.temp.geo.entity.Address;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
 /*
-* 국토교통부 Open API
-* https://www.vworld.kr/dev/v4dv_geocoderguide2_s002.do
-* */
+ * 국토교통부 Open API
+ * https://www.vworld.kr/dev/v4dv_geocoderguide2_s002.do
+ * */
 @Component
 public class Coord2AddressClient {
     private static final String SERVICE_PARAM = "service";
@@ -23,8 +22,8 @@ public class Coord2AddressClient {
     private static final String REQUEST_PARAM_VALUE = "getAddress";
     private static final String KEY_PARAM = "key";
     private static final String POINT_PARAM = "point";
-    private static final String RESPONSE_TYPE_PARAM ="type";
-    private static final String RESPONSE_TYPE_PARAM_VALUE ="PARCEL";    //법정동
+    private static final String RESPONSE_TYPE_PARAM = "type";
+    private static final String RESPONSE_TYPE_PARAM_VALUE = "PARCEL";    //법정동
 
     @Value("${geo.vworld.coord2address-uri}")
     private String coord2AddressUri;
