+++ src/main/java/com/example/temp/user/service/impl/SaveUserAuthenticatedLocationServiceImpl.java
@@ -1,8 +1,7 @@
 package com.example.temp.user.service.impl;
 
-import com.example.temp.common.entity.Address;
-import com.example.temp.geo.entity.AuthenticatedLocation;
-import com.example.temp.geo.entity.AuthenticatedLocationRepository;
+import com.example.temp.geo.entity.Address;
+import com.example.temp.geo.entity.AddressRepository;
 import com.example.temp.user.domain.User;
 import com.example.temp.user.domain.UserAuthenticatedLocation;
 import com.example.temp.user.domain.UserAuthenticatedLocationRepository;
@@ -19,14 +18,13 @@ import org.springframework.transaction.annotation.Transactional;
 public class SaveUserAuthenticatedLocationServiceImpl implements SaveUserAuthenticatedLocationService {
     private final UserRepository userRepository;
     private final UserAuthenticatedLocationRepository userAuthenticatedLocationRepository;
-    private final AuthenticatedLocationRepository authenticatedLocationRepository;
+    private final AddressRepository addressRepository;
 
     @Override
     public void doService(Long userId, UserAuthenticatedLocationRequest request) {
         User user = userRepository.findByIdOrElseThrow(userId);
         Address address = request.toEntity();
-
-        AuthenticatedLocation firstAuthenticationLocation = authenticatedLocationRepository.save(new AuthenticatedLocation(address));
+        Address firstAuthenticationLocation = addressRepository.save(address);
 
         UserAuthenticatedLocation entity = UserAuthenticatedLocation.build(user, firstAuthenticationLocation);
         userAuthenticatedLocationRepository.save(entity);
